+++ app/scripts/lib/transaction/hooks/enforce-simulation-hook.test.ts
@@ -1,23 +1,15 @@
 import {
-  SimulationTokenStandard,
   TransactionContainerType,
-  TransactionControllerEstimateGasAction,
   TransactionMeta,
   TransactionStatus,
 } from '@metamask/transaction-controller';
 import { EnforceSimulationHook } from './enforce-simulation-hook';
 import { TransactionControllerInitMessenger } from '../../../controller-init/messengers/transaction-controller-messenger';
-import { Messenger } from '@metamask/base-controller';
-import { cloneDeep } from 'lodash';
-import { DELEGATOR_CONTRACTS } from '@metamask/delegation-deployments';
-import { Hex, remove0x } from '@metamask/utils';
+import { Hex } from '@metamask/utils';
 import { ORIGIN_METAMASK } from '@metamask/controller-utils';
-import { DelegationControllerSignDelegationAction } from '@metamask/delegation-controller';
+import { applyTransactionContainers } from '../containers/util';
 
-const ESTIMATE_GAS_MOCK = '0x123' as Hex;
-const TOKEN_MOCK = '0x4567890abcdef1234567890abcdef1234567890a' as Hex;
-const DELEGATION_SIGNATURE_MOCK = '0x456aaabbbcccdddeee123' as Hex;
-const TOKEN_ID_MOCK = '0xabcdef345712ad' as Hex;
+jest.mock('../containers/util');
 
 const BALANCE_CHANGE_MOCK = {
   difference: '0x1' as Hex,
@@ -49,250 +41,69 @@ const TRANSACTION_META_MOCK: TransactionMeta = {
 };
 
 describe('EnforceSimulationHook', () => {
-  let messenger: TransactionControllerInitMessenger;
+  let messenger = {} as TransactionControllerInitMessenger;
 
-  const estimateGasMock: jest.MockedFn<
-    TransactionControllerEstimateGasAction['handler']
-  > = jest.fn();
-
-  const signDelegationMock: jest.MockedFn<
-    DelegationControllerSignDelegationAction['handler']
-  > = jest.fn();
+  const applyTransactionContainersMock = jest.mocked(
+    applyTransactionContainers,
+  );
 
   beforeEach(() => {
     jest.resetAllMocks();
 
     process.env.ENABLE_ENFORCED_SIMULATIONS = true as never;
 
-    const baseMessenger = new Messenger<
-      | DelegationControllerSignDelegationAction
-      | TransactionControllerEstimateGasAction,
-      never
-    >();
+    applyTransactionContainersMock.mockResolvedValue({
+      updateTransaction: jest.fn(),
+    });
+  });
 
-    baseMessenger.registerActionHandler(
-      'TransactionController:estimateGas',
-      estimateGasMock,
-    );
+  it('applies enforced simulations container if after simulate hook', async () => {
+    const updateTransactionMock = jest.fn();
 
-    baseMessenger.registerActionHandler(
-      'DelegationController:signDelegation',
-      signDelegationMock,
-    );
-
-    messenger = baseMessenger.getRestricted({
-      name: 'TransactionControllerInitMessenger',
-      allowedActions: [
-        'DelegationController:signDelegation',
-        'TransactionController:estimateGas',
-      ],
-      allowedEvents: [],
+    applyTransactionContainersMock.mockResolvedValue({
+      updateTransaction: updateTransactionMock,
     });
 
-    estimateGasMock.mockResolvedValue({
-      gas: ESTIMATE_GAS_MOCK,
-      simulationFails: {
-        debug: {},
-      },
+    const hook = new EnforceSimulationHook({
+      messenger,
+    }).getAfterSimulateHook();
+
+    const { updateTransaction } =
+      (await hook({
+        transactionMeta: TRANSACTION_META_MOCK,
+      })) ?? {};
+
+    expect(updateTransaction).toBe(updateTransactionMock);
+
+    expect(applyTransactionContainersMock).toHaveBeenCalledWith(
+      expect.objectContaining({
+        isApproved: false,
+      }),
+    );
+  });
+
+  it('applies enforced simulations container if before sign hook', async () => {
+    const updateTransactionMock = jest.fn();
+
+    applyTransactionContainersMock.mockResolvedValue({
+      updateTransaction: updateTransactionMock,
     });
 
-    signDelegationMock.mockResolvedValue(DELEGATION_SIGNATURE_MOCK);
-  });
-
-  it('updates transaction target to delegation manager', async () => {
     const hook = new EnforceSimulationHook({
       messenger,
-    }).getAfterSimulateHook();
+    }).getBeforeSignHook();
 
     const { updateTransaction } =
       (await hook({
         transactionMeta: TRANSACTION_META_MOCK,
       })) ?? {};
 
-    const newTransaction = cloneDeep(TRANSACTION_META_MOCK);
-    updateTransaction?.(newTransaction);
+    expect(updateTransaction).toBe(updateTransactionMock);
 
-    expect(newTransaction.txParams.to).toBe(
-      DELEGATOR_CONTRACTS['1.3.0']['1'].DelegationManager,
-    );
-  });
-
-  it('updates transaction data to redeemDelegations call', async () => {
-    const hook = new EnforceSimulationHook({
-      messenger,
-    }).getAfterSimulateHook();
-
-    const { updateTransaction } =
-      (await hook({
-        transactionMeta: TRANSACTION_META_MOCK,
-      })) ?? {};
-
-    const newTransaction = cloneDeep(TRANSACTION_META_MOCK);
-    updateTransaction?.(newTransaction);
-
-    expect(newTransaction.txParams.data?.substring(0, 10)).toStrictEqual(
-      '0xcef6d209',
-    );
-  });
-
-  it('updates transaction value to zero', async () => {
-    const hook = new EnforceSimulationHook({
-      messenger,
-    }).getAfterSimulateHook();
-
-    const { updateTransaction } =
-      (await hook({
-        transactionMeta: TRANSACTION_META_MOCK,
-      })) ?? {};
-
-    const newTransaction = cloneDeep(TRANSACTION_META_MOCK);
-    updateTransaction?.(newTransaction);
-
-    expect(newTransaction.txParams.value).toBe('0x0');
-  });
-
-  it('adds container type to transaction', async () => {
-    const hook = new EnforceSimulationHook({
-      messenger,
-    }).getAfterSimulateHook();
-
-    const { updateTransaction } =
-      (await hook({
-        transactionMeta: TRANSACTION_META_MOCK,
-      })) ?? {};
-
-    const newTransaction = cloneDeep(TRANSACTION_META_MOCK);
-    updateTransaction?.(newTransaction);
-
-    expect(newTransaction.containerTypes).toStrictEqual([
-      TransactionContainerType.EnforcedSimulations,
-    ]);
-  });
-
-  it('includes native balance change caveat', async () => {
-    const hook = new EnforceSimulationHook({
-      messenger,
-    }).getAfterSimulateHook();
-
-    const { updateTransaction } =
-      (await hook({
-        transactionMeta: TRANSACTION_META_MOCK,
-      })) ?? {};
-
-    const newTransaction = cloneDeep(TRANSACTION_META_MOCK);
-    updateTransaction?.(newTransaction);
-
-    expect(newTransaction.txParams.data).toStrictEqual(
-      expect.stringContaining(
-        remove0x(
-          DELEGATOR_CONTRACTS['1.3.0']['1'].NativeBalanceChangeEnforcer,
-        ).toLowerCase(),
-      ),
-    );
-  });
-
-  it('includes erc20 token balance change caveat', async () => {
-    const hook = new EnforceSimulationHook({
-      messenger,
-    }).getAfterSimulateHook();
-
-    const { updateTransaction } =
-      (await hook({
-        transactionMeta: {
-          ...TRANSACTION_META_MOCK,
-          simulationData: {
-            tokenBalanceChanges: [
-              {
-                ...BALANCE_CHANGE_MOCK,
-                address: TOKEN_MOCK,
-                standard: SimulationTokenStandard.erc20,
-              },
-            ],
-          },
-        },
-      })) ?? {};
-
-    const newTransaction = cloneDeep(TRANSACTION_META_MOCK);
-    updateTransaction?.(newTransaction);
-
-    expect(newTransaction.txParams.data).toStrictEqual(
-      expect.stringContaining(
-        remove0x(
-          DELEGATOR_CONTRACTS['1.3.0']['1'].ERC20BalanceChangeEnforcer,
-        ).toLowerCase(),
-      ),
-    );
-  });
-
-  it('includes erc721 token balance change caveat', async () => {
-    const hook = new EnforceSimulationHook({
-      messenger,
-    }).getAfterSimulateHook();
-
-    const { updateTransaction } =
-      (await hook({
-        transactionMeta: {
-          ...TRANSACTION_META_MOCK,
-          simulationData: {
-            tokenBalanceChanges: [
-              {
-                ...BALANCE_CHANGE_MOCK,
-                address: TOKEN_MOCK,
-                id: TOKEN_ID_MOCK,
-                standard: SimulationTokenStandard.erc721,
-              },
-            ],
-          },
-        },
-      })) ?? {};
-
-    const newTransaction = cloneDeep(TRANSACTION_META_MOCK);
-    updateTransaction?.(newTransaction);
-
-    expect(newTransaction.txParams.data).toStrictEqual(
-      expect.stringContaining(
-        remove0x(
-          DELEGATOR_CONTRACTS['1.3.0']['1'].ERC721BalanceChangeEnforcer,
-        ).toLowerCase(),
-      ),
-    );
-  });
-
-  it('includes erc1155 token balance change caveat', async () => {
-    const hook = new EnforceSimulationHook({
-      messenger,
-    }).getAfterSimulateHook();
-
-    const { updateTransaction } =
-      (await hook({
-        transactionMeta: {
-          ...TRANSACTION_META_MOCK,
-          simulationData: {
-            tokenBalanceChanges: [
-              {
-                ...BALANCE_CHANGE_MOCK,
-                address: TOKEN_MOCK,
-                id: TOKEN_ID_MOCK,
-                standard: SimulationTokenStandard.erc1155,
-              },
-            ],
-          },
-        },
-      })) ?? {};
-
-    const newTransaction = cloneDeep(TRANSACTION_META_MOCK);
-    updateTransaction?.(newTransaction);
-
-    expect(newTransaction.txParams.data).toStrictEqual(
-      expect.stringContaining(
-        remove0x(
-          DELEGATOR_CONTRACTS['1.3.0']['1'].ERC1155BalanceChangeEnforcer,
-        ).toLowerCase(),
-      ),
-    );
-
-    expect(newTransaction.txParams.data).toStrictEqual(
-      expect.stringContaining(remove0x(TOKEN_ID_MOCK).toLowerCase()),
+    expect(applyTransactionContainersMock).toHaveBeenCalledWith(
+      expect.objectContaining({
+        isApproved: true,
+      }),
     );
   });
 
@@ -359,40 +170,6 @@ describe('EnforceSimulationHook', () => {
         })) ?? {};
 
       expect(updateTransaction).toBeUndefined();
-    });
-  });
-
-  describe('if before sign hook', () => {
-    it('signs delegation', async () => {
-      const hook = new EnforceSimulationHook({
-        messenger,
-      }).getBeforeSignHook();
-
-      const { updateTransaction } =
-        (await hook({
-          transactionMeta: TRANSACTION_META_MOCK,
-        })) ?? {};
-
-      const newTransaction = cloneDeep(TRANSACTION_META_MOCK);
-      updateTransaction?.(newTransaction);
-
-      expect(newTransaction.txParams.data).toStrictEqual(
-        expect.stringContaining(
-          remove0x(DELEGATION_SIGNATURE_MOCK).toLowerCase(),
-        ),
-      );
-    });
-
-    it('does not estimate gas', async () => {
-      const hook = new EnforceSimulationHook({
-        messenger,
-      }).getBeforeSignHook();
-
-      await hook({
-        transactionMeta: TRANSACTION_META_MOCK,
-      });
-
-      expect(estimateGasMock).not.toHaveBeenCalled();
     });
   });
 });
